move project Overview classes to overview package
