+++ modules/features/project/project/src/main/res/navigation/project_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/overviewDest"
-        android:name="lu.aqu.projper.project.OverviewFragment"
+        android:name="lu.aqu.projper.project.overview.OverviewFragment"
         tools:layout="@layout/fragment_overview" />
 
 </navigation>
